OpenAI-wrapper/AzureTask2/function_app.py
import azure.functions as func
import logging
import openai_client

app = func.FunctionApp(http_auth_level=func.AuthLevel.FUNCTION)

@app.route(route="http_trigger")
def http_trigger(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request.')

    query = req.params.get('query')

    if query:
        llm_response = openai_client.call_openai_api(query)
        return func.HttpResponse(f"Hello, this is a LLM response: {llm_response} ", status_code=200)
    else:
        return func.HttpResponse(
             "This HTTP triggered function executed successfully. Pass a name in the query string or in the request body for a personalized response.",
             status_code=200
        )